modif %p

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ocviller <ocviller@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/11 20:33:23 by ocviller          #+#    #+#             */
-/*   Updated: 2025/06/04 15:38:38 by ocviller         ###   ########.fr       */
+/*   Updated: 2025/06/05 12:25:29 by ocviller         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int	ft_printstr(char *str);
 int	ft_printuns(unsigned int nb);
 int	ft_printhex(unsigned int nb);
 int	ft_printhex_up(unsigned int nb);
-int	ft_putnbr_base(long unsigned int nb, char *base);
+int	ft_putnbr_base(unsigned long long nb, char *base);
 int	ft_printmem(void *ptr);
 int	ft_print_pointer(void *ptr);
